+++ sudoc/image_management/include/black_n_white.h
