+++ sudoc/image_management/include/contrast_light.h
