+++ sudoc/image_management/include/crop.h
