+++ sudoc/image_management/include/crop_ia.h
