+++ sudoc/image_management/include/gaussian.h
