+++ sudoc/image_management/include/grayscale.h
