+++ sudoc/image_management/include/resize.h
