+++ sudoc/image_management/include/sobel.h
