+++ sudoc/image_management/include/to_matrix.h
@@ -2,7 +2,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <err.h>
-#include "../../libraries/matrix_lib/matrix.h"
+#include "../../libraries/matrix_lib/include/matrix.h"
 #include <math.h>
 #include <SDL2/SDL.h>
 #include <SDL2/SDL_image.h>
