+++ sudoc/image_management/libs/black_n_white.c
