+++ sudoc/image_management/libs/cell.c
