+++ sudoc/image_management/libs/contrast_light.c
