+++ sudoc/image_management/libs/crop.c
