+++ sudoc/image_management/libs/crop_ia.c
