+++ sudoc/image_management/libs/gaussian.c
