+++ sudoc/image_management/libs/grayscale.c
