+++ sudoc/image_management/libs/hough_transform.c
@@ -116,24 +116,6 @@ SDL_Surface *CV_IMG_TO_SURFACE(const Image *image)
 }
 
 
-
-// Loads an image in a surface.
-// The format of the surface is SDL_PIXELFORMAT_RGB888.
-//
-// path: Path of the image.
-SDL_Surface* Load_image(const char* path)
-{
-    SDL_Surface* surface = IMG_Load(path);
-    if (surface == NULL)
-        errx(EXIT_FAILURE, "%s", SDL_GetError());
-    SDL_Surface* res = SDL_ConvertSurfaceFormat(surface,SDL_PIXELFORMAT_RGB888,0);
-    if (res == NULL)
-        errx(EXIT_FAILURE, "%s", SDL_GetError());
-    SDL_FreeSurface(surface);
-    return res;
-}
-
-
 /// @brief Find Lines in an image using the Hough Transform algorithm.
 /// @param src The source image
 /// @param threshold The threshold value, representing the minimum number of intersections to detect a line
