+++ sudoc/image_management/libs/resize.c
