+++ sudoc/image_management/libs/sobel.c
