+++ sudoc/image_management/libs/to_matrix.c
